Refactor to composable
